chore: split CompositionTarget.skia.cs
